Move Desc to its own file, add helper function to get similar descriptions
