+++ examples/animation.py
@@ -16,8 +16,8 @@ import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.animation import FuncAnimation
 
-from data_prototype.containers import Desc
 from data_prototype.conversion_edge import Graph
+from data_prototype.description import Desc
 
 from data_prototype.conversion_node import FunctionConversionNode
